submit_jobs.py as module
Put scipt submit_jobs.py into function to make it testable with pytest. Most part of the diff comes from the indentation.
Function arguments are the command line options. The argument parsing is done when executing the module and uses the same default values. All calls to the namespace 'options' are replaced by the function arguments itself.